+++ src/components/appointments/Appointment.js
@@ -1,43 +1,35 @@
-import React, { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { React } from 'react';
 // import { useParams } from 'react-router-dom';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
-import { allSellers } from '../../redux/sellers/sellers';
+
 import Cities from '../cities/cities';
+import Sellers from '../sellers/sellers';
 
-const Appointment = () => {
-  const dispatch = useDispatch();
-  const sellers = useSelector((state) => state.sellers);
-  console.log(sellers);
-  // console.log(sellers);
-  // const { userId, carId } = useParams();
-  // console.log('UserId '+ userId);
-  // console.log('carId is ' + carId);
-  useEffect(() => {
-    dispatch(allSellers());
-  }, [dispatch]);
-
-  return (
+const Appointment = () =>
+// console.log(sellers);
+// const { userId, carId } = useParams();
+// console.log('UserId '+ userId);
+// console.log('carId is ' + carId);
+  // eslint-disable-next-line implicit-arrow-linebreak
+  (
     <div className="user-container" data-testid="user-a">
       <h1>Appointment Component</h1>
       <Form>
         <Cities />
+        <Sellers />
         <Form.Group className="mb-3" controlId="formBasicEmail">
           <Form.Label>Duration</Form.Label>
           <Form.Control type="email" placeholder="Enter duration" />
-          {/* <Form.Text className="text-muted">
-            We will never share your email with anyone else.
-          </Form.Text> */}
         </Form.Group>
 
         <Form.Group className="mb-3" controlId="formBasicPassword">
-          <Form.Label>Password</Form.Label>
-          <Form.Control type="password" placeholder="Password" />
+          <Form.Label>Enter Branch</Form.Label>
+          <Form.Control type="password" placeholder="Enter Branch" />
         </Form.Group>
 
         <Form.Group className="mb-3" controlId="formBasicCheckbox">
-          <Form.Check type="checkbox" label="Check me out" />
+          <Form.Check type="checkbox" label="Default Locat Branch" />
         </Form.Group>
 
         <Button variant="primary" type="submit">
@@ -46,6 +38,4 @@ const Appointment = () => {
       </Form>
     </div>
   );
-};
-
 export default Appointment;
